Major Refactoring part 1
This is the first part of a major refactoring of code. This update adds an event system for which you register a callback. This will more easily allow developers to take their own desired action based on the different states of the controller. The end result is code that is much easier to read and much easier to change or adapt to different form factors/pcbs.

- SDK configurator options have been revised
- Added option to disable EFM8UB1 USB companion
- Added option to disable I2C handling by HOJA (if you need to use that for your own purposes)
- Added option to disable RGB LED utility
- Added some modularity to the battery utility to allow for other PMICs to be integrated in the future.
- Renamed and changed how the button/analog callback functions work. The global variables are passed through to the callback functions as pointers to allow you to act on them in a clean manner. Rather than needing to hunt down the exact name of the global structs for buttons/analog inputs, they are conveniently passed as a pointer. 

New utilities are being added! 
- SOCD cleaner functions planned
- Snapback reduction functions planned
- general utility function file added. First function in this lineup is one that allows you to grab a bit from a 32 bit unsigned int.

The example code is still needing revision to get up to speed with all of these new features. I'm working hard on adding documentation to our wiki at https://wiki.handheldlegend.com as soon as possible to get more people up and running with this library!